requirements.txt
certifi==2020.11.8
chardet==3.0.4
Django==2.2
django-filter==2.4.0
djangorestframework==3.12.2
idna==2.10
Markdown==3.3.3
Pillow==8.0.1
pytz==2020.4
requests==2.25.0
sqlparse==0.4.1
urllib3==1.26.2
uuid==1.30
certifi==2020.11.8
chardet==3.0.4
Django==2.2
django-filter==2.4.0
djangorestframework==3.12.2
gunicorn==20.0.4
idna==2.10
Markdown==3.3.3
Pillow==8.0.1
pytz==2020.4
requests==2.25.0
sqlparse==0.4.1
urllib3==1.26.2
uuid==1.30
django-heroku
whitenoise
